requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
asyncpg==0.20.1
attrs==19.3.0
chardet==3.0.4
click==7.1.2
discord.py==1.3.4
idna==2.10
multidict==4.7.6
numpy==1.19.1
typing-extensions==3.7.4.2
websockets==8.1
yarl==1.5.0
